Add json file chek & updates
Adds automatic exchange rates JSON file updates on server start and check if it is up to date.
Checks if the JSON file exists and creates a new one if not.
Checks if the current JSON file is of the same date as UTC not date and updates it if not.